Fix shutdown hook to handle failed startup correctly
